+++ models/Page.php
@@ -2,6 +2,7 @@
 
 namespace humhub\modules\custom_pages\models;
 
+use humhub\libs\Html;
 use humhub\modules\content\models\Content;
 use humhub\modules\custom_pages\helpers\Url;
 use humhub\modules\custom_pages\models\forms\SettingsForm;
@@ -186,6 +187,10 @@ class Page extends CustomContentContainer
      */
     public function getPageContent()
     {
+        if ($this->type == HtmlType::ID) {
+            return preg_replace('/(<script.*?) nonce(="")?(.*?>)/i', '$1 ' . Html::nonce() . '$3', $this->page_content);
+        }
+
         return $this->page_content;
     }